+++ src/majordomo_llm/providers/anthropic.py
@@ -169,6 +169,7 @@ class Anthropic(LLM):
             output_cost=output_cost,
             total_cost=total_cost,
             response_time=execution_time,
+            deprecation_warning=self.deprecation_warning,
         )
 
     async def get_response_stream(