+++ src/majordomo_llm/providers/cohere.py
@@ -170,6 +170,7 @@ class Cohere(LLM):
             output_cost=output_cost,
             total_cost=total_cost,
             response_time=execution_time,
+            deprecation_warning=self.deprecation_warning,
         )
 
     async def get_response_stream(